python/requirements.txt
numpy==1.18.1
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
